fixed kang dimention error

+++ userge/plugins/fun/kang.py
@@ -8,6 +8,7 @@
 #
 # All rights reserved.
 
+import io
 import os
 import random
 import urllib.request
@@ -169,7 +170,8 @@ async def kang_(message: Message):
                 await conv.get_response(mark_read=True)
         await message.edit(f"`Sticker kanged successfully!`\n"
                            f"Pack can be found [here](t.me/addstickers/{packname})")
-        os.remove(photo)
+        if os.path.exists(str(photo)):
+            os.remove(photo)
 
 
 @userge.on_cmd("stkrinfo", about={
@@ -204,18 +206,18 @@ async def sticker_pack_info_(message: Message):
     await message.edit(out_str)
 
 
-def resize_photo(photo: str) -> str:
+def resize_photo(photo: str) -> io.BytesIO:
     """ Resize the given photo to 512x512 """
     image = Image.open(photo)
-    scale = 512 / max(image.width, image.height)
+    maxsize = 512
+    scale = maxsize / max(image.width, image.height)
     new_size = (int(image.width*scale), int(image.height*scale))
-    image = image.resize(new_size, Image.ANTIALIAS)
+    image = image.resize(new_size, Image.LANCZOS)
+    resized_photo = io.BytesIO()
+    resized_photo.name = "sticker.png"
+    image.save(resized_photo, "PNG")
     os.remove(photo)
-    photo = os.path.join(Config.DOWN_PATH, "sticker.png")
-    if os.path.exists(photo):
-        os.remove(photo)
-    image.save(photo, "PNG")
-    return photo
+    return resized_photo
 
 
 KANGING_STR = (
